# lib/api-spec/openapi-contact-crm.yaml
openapi: 3.1.0
info:
  title: ClaimTagX Contact CRM
  version: 0.1.0
  description: |
    OpenAPI 3.1 fragment for ClaimTagX public Contact Us and Platform Contact
    Operations. Paths are relative to the same `/api` server as `openapi.yaml`.

    **Auth**
    - Public `/contact/*` form endpoints are unauthenticated.
    - Inbound webhooks use `X-Webhook-Secret`.
    - Platform (and Handler) routes use first-party ClaimTagX auth: the
      `ctx_auth_session` cookie (an opaque session/refresh token) **or** an
      opaque `Authorization: Bearer` token. No hosted identity provider is used.
      Sign in via `POST /platform/auth/login` (email + password); MFA, password
      reset, invitation acceptance and first-owner bootstrap are also
      first-party endpoints under `/platform/auth/*`.

    **Permissions**
    Platform operations declare `x-permission` matching `PLATFORM_PERMISSIONS`
    in `artifacts/api-server/src/lib/crm/rbac.ts`. `GET /platform/me` requires
    an authenticated staff session but no specific permission.

    **Status**
    Live handlers are `x-status: implemented`. Remaining stubs are
    `x-status: planned`.
servers:
  - url: /api
    description: Base API path (mounted in artifacts/api-server/src/app.ts)
...
paths:
...
  /platform/contact/effects:
    get:
      operationId: listPlatformJobEffects
      tags: [platform-workspace]
      summary: Inspect job effects
      x-status: implemented
      x-permission: jobs.inspect
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Effect rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmJobEffectsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /platform/contact/effects/{key}:
    get:
      operationId: getPlatformJobEffect
      tags: [platform-workspace]
      summary: Load one job effect by idempotency key
      x-status: implemented
      x-permission: jobs.inspect
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Effect row
          content:
            application/json:
              schema:
                type: object
                required: [effect]
                properties:
                  effect:
                    type: object
                    additionalProperties: true
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /platform/contact/effects/{key}/reconcile:
    post:
      operationId: reconcilePlatformJobEffect
      tags: [platform-workspace]
      summary: Reconcile an uncertain outbound email effect
      x-status: implemented
      x-permission: jobs.replay
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reconciliation outcome
          content:
            application/json:
              schema:
                type: object
                required: [key, outcome]
                properties:
                  key: { type: string }
                  outcome: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /platform/contact/effects/{key}/replay:
    post:
      operationId: replayPlatformJobEffect
      tags: [platform-workspace]
      summary: Authorized terminal/uncertain effect replay
      x-status: implemented
      x-permission: jobs.replay
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Replayed
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ctx_auth_session
      description: >
        First-party opaque session cookie (`ctx_auth_session`). HttpOnly, host-only
        (no Domain), `SameSite=Lax`, `path=/`; Secure in production. claimtagx.com
        and api.claimtagx.com are schemeful same-site, so Lax covers SPA→API
        credentialed calls. Its value is an opaque refresh token backed by the
        first-party auth platform; only its hash is persisted server-side.
        Cookie-bearing writes are additionally protected by an Origin allow-list
        (CSRF) check.
    bearerToken:
      type: http
      scheme: bearer
      description: >
        First-party opaque access token (or refresh token) via
        `Authorization: Bearer`. Used by non-browser/API clients. Tokens are
        opaque (not JWTs); only their hashes are stored.
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared secret for inbound email and scheduling webhooks.
...
  responses:
    Unauthorized:
      description: Missing or invalid staff session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContactApiError"
    Forbidden:
      description: "Authenticated but missing `x-permission`"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContactApiError"
...
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContactApiError"
...
    Conflict:
      description: Optimistic concurrency conflict or invalid state transition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContactApiError"
...
    RateLimited:
      description: Admin or public rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContactApiError"
  schemas:
    ContactApiError:
      type: object
      properties:
        error:
          type: string
        correlationId:
          type: string
          description: Present on public submit errors.
      required:
        - error
...
    CrmJobEffectsListResponse:
      type: object
      required: [effects]
      properties:
        effects:
          type: array
          items:
            type: object
            additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
...